requirements.txt
django==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==22.1